Optimized JsLookupResult.asOpt for failure case
Currently, the asOpt method in JsReadable is implemented in terms of validate.  The validate method in JsLookupResult, in the JsUndefined case, ends up serializing the parent object referenced in the message to a string, which can be quite an expensive operation depending on the object's size; this message is then discarded.

This change re-implements asOpt in JsLookupResult in terms of JsLookupResult's toOption method.